refactor(profile): extract ProfileTabBar to stay under file-size ceiling
Reconciling the stale-while-revalidate cache bars (#5279/#5282) with the
list-driven own-profile Collabs tab (#5213) on one branch pushed
profile_grid.dart to 835 lines, over the 800-line ceiling (#4339). Move
the sticky tab bar (ProfileTabBar plus its private _ProfileTab and
_SliverAppBarDelegate) into profile_tab_bar.dart. No behavior change;
profile_grid.dart drops to 609 lines.
